% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.autoPlotName}
\alias{ww.autoPlotName}
\title{ww.autoPlotName}
\usage{
ww.autoPlotName(name = NULL)
}
\arguments{
\item{name}{Manually name your plot}
}
\description{
Internal function. Creates automatic plot and file-names.
}
\examples{
ww.autoPlotName()
}
