% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.FnP_parser}
\alias{ww.FnP_parser}
\title{ww.FnP_parser}
\usage{
ww.FnP_parser(fname, ext_wo_dot)
}
\arguments{
\item{fname}{Name of the file}

\item{ext_wo_dot}{File extension without separating dot.}
}
\description{
Internal Function. Parses the full path from the filename & location of the file.
}
\examples{
ww.FnP_parser(fname = 'myplot', ext_wo_dot = "jpg")
}
