% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.MarkDown_ImgLink_formatter}
\alias{ww.MarkDown_ImgLink_formatter}
\title{ww.MarkDown_ImgLink_formatter}
\usage{
ww.MarkDown_ImgLink_formatter(...)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}
}
\description{
Format a markdown image reference (link) from the file path to the file. It can parse the file path, if you pass it in separate variables and strings. E.g. ww.MarkDown_ImgLink_formatter(Directory, "MyImage.png").
}
\examples{
ww.MarkDown_ImgLink_formatter (... =  )
}
