% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.ttl_field}
\alias{ww.ttl_field}
\title{ww.ttl_field}
\usage{
ww.ttl_field(flname = fname)
}
\arguments{
\item{flname}{Name of the plot}
}
\description{
Internal function. Creates the string written into the PDF files "Title' (metadata) field.
}
\examples{
ww.ttl_field()
}
